crop_boxes/requirements.txt
botocore==1.29.76
jmespath==1.0.1
pdf2image==1.16.3
Pillow==9.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.14
aioboto3==11.1.0
pytesseract==0.3.10